+++ src/components/BurgerIngredientsConstructorWrapper/burger-ingredients-constructor-wrapper.module.css
@@ -1,0 +1,6 @@
+.twoRowedElements{
+    display: flex;
+    flex-direction: row;
+    justify-content: space-evenly;
+    align-items: flex-start;
+}
