+++ src/utils/common-data-from-server-context.js
@@ -1,0 +1,2 @@
+import React from 'react';
+export const CommonDataFromServerContext = React.createContext();
